+++ tests/testthat/testLibDynam/R/a.r
@@ -1,0 +1,9 @@
+a <- 1
+
+nulltest3 <- function() {
+  .Call("null_test", PACKAGE = "testDllLoad")
+}
+
+nulltest4 <- function() {
+  .Call(null_test2)
+}